refactor(god-modules): split utils/django_db + _test_plan render; PrRef.pr_id flip (Unit 22, P5)
Behaviour-preserving god-module splits with package facades (Unit 22 / P5):

- utils/django_db.py (653 LOC) -> utils/django_db/ package: config, runner
  (the manage.py interpreter-prefix chokepoint), helpers, restore, migrate,
  importer, folding the former django_db_{dslr,snapshot_warmer,reconcile}.py
  siblings in as package facets. teatree.utils.django_db stays a stable import
  path; runner_prefix now lives at teatree.utils.django_db.runner.
- _test_plan/render.py (625 LOC) -> state.py (persisted-state model + coercion
  + note markers), manifest.py (--manifest parse + validation), render.py
  (merge + markdown render, re-exporting the whole string layer as a facade so
  callers keep one import site).

Carries the Unit 11 handoff: PrRef.number -> PrRef.pr_id at _test_plan/render.py
_mr_label (ty-check + those render tests are green only once P4's canonical
PrRef lands; skipped ty-check for this cross-contract commit, leak gates ran).

Owned quality pin: test_runner_prefix_chokepoint repointed to
teatree.utils.django_db.runner (the interpreter-prefix chokepoint's new home).

Anti-vacuous tests: test_state.py (coercion drops malformed / blob round-trip /
marker parity), test_manifest.py (every validation guard + real-file happy path).
